requirements.txt
matplotlib==3.7.2
numpy==1.25.2
pandas==2.2.1
Pillow==10.2.0
streamlit==1.31.1
streamlit_drawable_canvas==0.9.3
torch==2.0.1
torchvision==0.15.2
xgboost==2.0.3